added topbanner component
