+++ package.json
@@ -22,7 +22,6 @@
     "next": "^12.0.9",
     "react": "17.0.2",
     "react-audio-player": "^0.17.0",
-    "react-bootstrap": "^1.6.1",
     "react-dom": "17.0.2",
     "react-scroll": "^1.8.4"
   },
