+++ src/pages/getinvolved.jsx
@@ -1,17 +1,7 @@
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
-import {
-  Box,
-  Button,
-  Card,
-  CardMedia,
-  IconButton,
-  Slide,
-  Typography,
-} from '@mui/material'
+import { Box, Button, Card, CardMedia } from '@mui/material'
 import { styled } from '@mui/material'
 import Head from 'next/head'
 import Image from 'next/image'
-import React, { useEffect, useState } from 'react'
 import { Link as ScrollElement } from 'react-scroll'
 
 import logo3 from '../../public/images/ccollege.png'
@@ -26,81 +16,12 @@ import roadmap from '../../public/images/roadmap.png'
 import topbanner from '../../public/images/srkw2-25.jpg'
 import logo1 from '../../public/images/twt.png'
 import logo11 from '../../public/images/ws_logo.png'
+import TopBanner from '../components/TopBanner'
 import getinvolvedStyles from '../styles/getinvolved.module.css'
-
-const TopScreen = styled('div')(({ theme }) => ({
-  minHeight: '90vh',
-  maxHeight: '90vh',
-  maxWidth: '100vw',
-  backgroundImage: `url(${topbanner.src})`,
-  backgroundPosition: 'center center',
-  backgroundSize: 'cover',
-  backgroundRepeat: 'no-repeat',
-  [theme.breakpoints.down('sm')]: {
-    minHeight: '68vh',
-    maxHeight: '68vh',
-    maxWidth: '100vw',
-  },
-}))
-
-const TitleScreen = styled('div')(({ theme }) => ({
-  position: 'relative',
-  minHeight: '85vh',
-  maxHeight: '85vh',
-  maxWidth: '100vw',
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-  justifyContent: 'center',
-  color: '#fff',
-  [theme.breakpoints.down('sm')]: {
-    minHeight: '68vh',
-    maxHeight: '68vh',
-    maxWidth: '100vw',
-    paddingBottom: '200px',
-  },
-}))
-
-const ScrollDownButton = styled(IconButton)(({ theme }) => ({
-  position: 'absolute',
-  bottom: '0',
-  padding: '0',
-  display: 'flex',
-  flexDirection: 'column',
-  alignItems: 'center',
-  justifyContent: 'center',
-  height: '15px',
-  transition: 'all 0.5s ease-in-out',
-  '&:hover': {
-    transform: 'translateY(5px)',
-  },
-  [theme.breakpoints.down('sm')]: {
-    display: 'none',
-  },
-}))
-
-const PageDesc = styled('div')(({ theme }) => ({
-  position: 'absolute',
-  bottom: '0',
-  left: '10px',
-  backgroundColor: '#080d26',
-  width: '40vw',
-  maxWidth: '450px',
-  padding: '20px',
-  margin: '10px 25px',
-  [theme.breakpoints.down('sm')]: {
-    left: '0',
-    margin: '0',
-    width: '100%',
-    maxWidth: '100vw',
-    margin: '0',
-  },
-}))
 
 const ScrollLink = styled(ScrollElement)({
   fontSize: '20px',
   color: '#1b2b7b',
-  textDecoration: 'none !important',
   cursor: 'pointer',
   '&:hover': {
     color: '#080d26',
@@ -108,55 +29,24 @@ const ScrollLink = styled(ScrollElement)({
 })
 
 export const Getinvolved = () => {
-  const [checked, setChecked] = useState(false)
   const InviteLinks = [
     { name: 'Volunteer', id: 'volunteer' },
     { name: 'For Developers', id: 'for-developers' },
     { name: 'Donate', id: 'donate' },
   ]
 
-  useEffect(() => {
-    setChecked(true)
-  }, [])
-
   return (
     <>
       <Head>
         <title>Orcasound</title>
       </Head>
-      <TopScreen>
-        <Slide
-          in={checked}
-          direction="up"
-          {...(checked ? { timeout: 1000 } : {})}
-        >
-          <TitleScreen>
-            <Typography
-              variant="h1"
-              sx={{ fontSize: '10vw', marginBottom: '2vw', fontWeight: '500' }}
-            >
-              Get Involved
-            </Typography>
-            <ScrollElement to="invite-links" smooth={true} spy={true}>
-              <ScrollDownButton>
-                <ExpandMoreIcon sx={{ fontSize: '5vw', color: '#ffffff' }} />
-                <ExpandMoreIcon
-                  sx={{
-                    fontSize: '5vw',
-                    color: '#ffffff',
-                    transform: 'translateY(-3.5vw)',
-                  }}
-                />
-              </ScrollDownButton>
-            </ScrollElement>
-            <PageDesc>
-              {`There are many ways to help in the recovering of marine life,
-              especially for the Souther Resident Killer Whales that call the
-              Salish Sea home. Check out the ways you can help below!`}
-            </PageDesc>
-          </TitleScreen>
-        </Slide>
-      </TopScreen>
+      <TopBanner
+        bannerImg={topbanner}
+        pageTitle={`Get Involved`}
+        pageDesc={`There are many ways to help in the recovering of marine life,
+          especially for the Souther Resident Killer Whales that call the
+          Salish Sea home. Check out the ways you can help below!`}
+      />
 
       <Box
         id="invite-links"
@@ -183,13 +73,13 @@ export const Getinvolved = () => {
             <div className={getinvolvedStyles.rows}>
               <h2>Citizen Scientist</h2>
               <p>
-                {`First and foremost, you can volunter as a citizen scientist. Not
+                First and foremost, you can volunter as a citizen scientist. Not
                 only an you listen for whales to marvel at the symphony of
-                ssounds they make.`}
+                ssounds they make.
                 <br></br>
-                {`but also you can listen 'for'
-                whales-- helping monitor their habitat and notifying the network
-                when you hear them or a noise that could endager them.`}
+                but also you can listen 'for' whales-- helping monitor their
+                habitat and notifying the network when you hear them or a noise
+                that could endager them.
               </p>
             </div>
           </div>
@@ -197,13 +87,13 @@ export const Getinvolved = () => {
           <div className={getinvolvedStyles.rows}>
             <h2>In Person</h2>
             <p>
-              {`You can also volunteer in-person with any of them current
+              You can also volunteer in-person with any of them current
               Orcasound network members, or with a new organizationin your
-              neighborhood that you convince to become a new member`}
+              neighborhood that you convince to become a new member
               <br></br>
-              {`Volunteer opportunities can include helping deploy or fix
+              Volunteer opportunities can include helping deploy or fix
               hydrophones, teaching groups how to 'listen for whales,' or
-              helping create new educational or outreach projects.`}
+              helping create new educational or outreach projects.
             </p>
           </div>
 
@@ -225,12 +115,12 @@ export const Getinvolved = () => {
           <div className={getinvolvedStyles.car}>
             <h2> For Developers</h2>
             <p>
-              {`A growing team of volunteer developers and talented consultants
+              A growing team of volunteer developers and talented consultants
               are building and improving Orcasound. Some are developing the new
               open-source software that captures and conveys ocean sounds.
               Others are pioneering the DIY , low cost hydrophone and
               computerhardware that allows humans to listen to marine
-              soundscapes in more and more places`}
+              soundscapes in more and more places
             </p>
           </div>
         </div>
@@ -246,10 +136,10 @@ export const Getinvolved = () => {
         </div>
 
         <p className={getinvolvedStyles.paragraph}>
-          {` Our crown jewel is the Orcasound Web App-- a suite of new cloud and
+          Our crown jewel is the Orcasound Web App-- a suite of new cloud and
           browser based ways for citizens scientist and artifical intelligence
           to listen for whales in real time.As of april 2020 this is the
-          Orcasound tech Stack`}
+          Orcasound tech Stack
         </p>
 
         <ul className={getinvolvedStyles.list}>
@@ -272,14 +162,14 @@ export const Getinvolved = () => {
           <br></br>
         </div>
         <p className={getinvolvedStyles.textroadmap}>
-          {` If you're based in the Pacific Northwest you can work with Orcasound
+          If you're based in the Pacific Northwest you can work with Orcasound
           in-person at a hackathon (see the Orcasound projects at democracy lab)
           No matter where you are, you can join Orcasound Slack, check out our
           Github repository and Trello boards, subscribe to the Orcasound dev
           email distribution list, and finish a place to contribute your
           talents. We hope you will share your expertise and innovations with
           us, and maybe even earn your way into the Orcasound HacHacker hall of
-          Fame`}
+          Fame
         </p>
         <div className={getinvolvedStyles.button}>
           <Button variant="outlined">LEARN MORE ABOUT VOLUNTEERING!</Button>
@@ -288,54 +178,54 @@ export const Getinvolved = () => {
           Memorandum Of Agreements
         </h2>
         <p className={getinvolvedStyles.textmemorandum}>
-          {` The real time audio streams, citizen science projects, educational
+          The real time audio streams, citizen science projects, educational
           material and outreach projects of Orcasound are brought to you by the
           current network member, listed below who have e-signed the 2016-2020
           Memorandum of Agreements(MOA)Any organization or individual is welcome
           to join the network(for free!)either as the host of an hydrophone
           node, a researcher or citizen scientist an educator/activist or
-          general volunteer.`}
+          general volunteer.
           <br></br>
-          {` If you are an individual wanting to volunteer , collaborate or donate,
+          If you are an individual wanting to volunteer , collaborate or donate,
           check out the many ways you can support Orcasound.Everyone can listen
-          for whales, and learn the diverse sounds of Salish Sea.`}
+          for whales, and learn the diverse sounds of Salish Sea.
           <br></br>
-          {`If you are an organization wanting to join the network as the host of
+          If you are an organization wanting to join the network as the host of
           a new hydrophone node, an educational/outreacg node, or both -- just
           read the history, mission and vision of the netowrk e-sign the MOA and
           then email info@orcasound.net to begin collaborating.There are no
-          membership fees-- just benefits roles and responsibilities.`}
+          membership fees-- just benefits roles and responsibilities.
         </p>
         <h2 className={getinvolvedStyles.donate} id="donate">
           {' '}
           Donate
         </h2>
         <p className={getinvolvedStyles.donatetext}>
-          {`Help us and our Orcasound network members by making charitable
+          Help us and our Orcasound network members by making charitable
           contribution to our partners, many of whom are 501(c)3 organizations
           Check out the link below to help strengthen and grow our network,
           while supporting our on-going conservation, research, and educational
-          efforts.`}
+          efforts.
           <br></br>
-          {`You can also directly support the many dedicated volunteers who help
+          You can also directly support the many dedicated volunteers who help
           Orcasound keep running and improve over time. Take a look at our
           'Hacker hall of fame ' and our Github repositories and consider
-          sponsoring the work of our most-dedicated contributors.`}
+          sponsoring the work of our most-dedicated contributors.
         </p>
         <div className={getinvolvedStyles.button}>
           <Button variant="outlined"> DONATE NOW</Button>
         </div>
 
         <div className={getinvolvedStyles.logos}>
-          <Image src={logo1} width={150} height={150} alt="donate-here" />
-          <Image src={logo2} width={150} height={150} alt="donate-here" />
-          <Image src={logo3} width={150} height={150} alt="donate-here" />
-          <Image src={logo4} width={150} height={150} alt="donate-here" />
-          <Image src={logo5} width={150} height={150} alt="donate-here" />
-          <Image src={logo6} width={150} height={150} alt="donate-here" />
-          <Image src={logo8} width={150} height={150} alt="donate-here" />
-          <Image src={logo9} width={150} height={150} alt="donate-here" />
-          <Image src={logo11} width={150} height={150} alt="donate-here" />
+          <Image src={logo1} width={150} height={150} />
+          <Image src={logo2} width={150} height={150} />
+          <Image src={logo3} width={150} height={150} />
+          <Image src={logo4} width={150} height={150} />
+          <Image src={logo5} width={150} height={150} />
+          <Image src={logo6} width={150} height={150} />
+          <Image src={logo8} width={150} height={150} />
+          <Image src={logo9} width={150} height={150} />
+          <Image src={logo11} width={150} height={150} />
         </div>
       </div>
     </>
